+++ src/globalState/plusScoreRuleAtom.ts
@@ -1,6 +1,6 @@
 import { atomWithStorage } from "jotai/utils";
 
-type PlusScoreRule = 25000 | 30000;
+export type PlusScoreRule = 25000 | 30000;
 
 export const plusScoreRuleAtom = atomWithStorage<PlusScoreRule | null>(
   "plusScoreRule",
